@@ -1,11 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\FBT.Generator\FBT.csproj" />
-  </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\FBT.Generator\FBT.csproj"/>
+    </ItemGroup>
 
 </Project>
