@@ -1,27 +1,23 @@
 ﻿using FBT.TypeData;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace FBT.Generator
+namespace FBT.Generator;
+
+public abstract class FBTCodeGenerator
 {
-    public abstract class FBTCodeGenerator
-    {
-        public FBTCodeGenerator(string p_Name, string p_Path = "./")
-        {
-            this.OutPath = p_Path;
-        }
+	public string OutPath;
 
-        public string OutPath;
+	public FBTCodeGenerator(string p_Name, string p_Path = "./")
+	{
+		OutPath = p_Path;
+	}
 
-        public virtual void PreBuild()
-        {
-        }
+	public virtual void PreBuild()
+	{
+	}
 
-        public virtual void PostBuild()
-        {
-        }
+	public virtual void PostBuild()
+	{
+	}
 
-        public abstract void GenerateUnit(TypeUnit p_Unit, string p_Name);
-    }
+	public abstract void GenerateUnit(TypeUnit p_Unit, string p_Name);
 }
