@@ -1,30 +1,25 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
-namespace FBT.Module
+namespace FBT.Module;
+
+public class ModuleManager
 {
-    public class ModuleManager
-    {
-        List<ModuleInfo> Modules = new List<ModuleInfo>( );
+	private static ModuleManager _Instance;
+	private readonly List<ModuleInfo> Modules = new();
+	public static ModuleManager Instance => _Instance ?? (_Instance = new ModuleManager());
 
 
-        public ModuleInfo GetModule( string p_Name )
-        {
-            var s_Module = this.Modules.FirstOrDefault( x => x.Name == p_Name );
+	public ModuleInfo GetModule(string p_Name)
+	{
+		var s_Module = Modules.FirstOrDefault(x => x.Name == p_Name);
 
-            if ( s_Module == null )
-            {
-                s_Module = new ModuleInfo( p_Name );
-                this.Modules.Add( s_Module );
-            }
+		if (s_Module == null)
+		{
+			s_Module = new ModuleInfo(p_Name);
+			Modules.Add(s_Module);
+		}
 
-            return s_Module;
-        }
-
-
-        static ModuleManager _Instance;
-        public static ModuleManager Instance => _Instance ?? ( _Instance = new ModuleManager( ) );
-    }
+		return s_Module;
+	}
 }
