@@ -1,26 +1,30 @@
 ﻿using FBT.TypeData.Base;
 using FBT.TypeData.Base.Attributes;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace FBT.TypeData.Member
+namespace FBT.TypeData.Member;
+
+public class TypeDataValueType
+	: TypeDataBase
 {
-    public class TypeDataValueType
-        : TypeDataBase
-    {
-        public string Protection = null;
+	public string Protection = null;
 
+	public uint? GetMessageType()
+	{
+		var s_NumeralAttribute = FindAttributeIgnoreCase("MessageType") as TypeNumeralAttribute;
 
-        public int GetSize( )
-        {
-            var s_NumeralAttribute = this.FindAttributeIgnoreCase( "size" ) as TypeNumeralAttribute;
+		if (s_NumeralAttribute == null)
+			return null;
 
-            if ( s_NumeralAttribute == null )
-                return -1;
+		return (uint)s_NumeralAttribute.Value;
+	}
 
-            return s_NumeralAttribute.Value;
-        }
+	public uint? GetMessageCategory()
+	{
+		var s_NumeralAttribute = FindAttributeIgnoreCase("MessageCategory") as TypeNumeralAttribute;
 
-    }
+		if (s_NumeralAttribute == null)
+			return null;
+
+		return (uint)s_NumeralAttribute.Value;
+	}
 }
