@@ -1,14 +1,14 @@
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
-//     ANTLR Version: 4.7.2
+//     ANTLR Version: 4.13.1
 //
 //     Changes to this file may cause incorrect behavior and will be lost if
 //     the code is regenerated.
 // </auto-generated>
 //------------------------------------------------------------------------------
 
-// Generated from G:\Users\txt\Source\Repos\FBTC1\FBT.Generator\Parser\Grammar\FBT.g4 by ANTLR 4.7.2
+// Generated from FBT.g4 by ANTLR 4.13.1
 
 // Unreachable code detected
 #pragma warning disable 0162
@@ -31,7 +31,8 @@ using ParserRuleContext = Antlr4.Runtime.ParserRuleContext;
 /// which can be extended to create a listener which only needs to handle a subset
 /// of the available methods.
 /// </summary>
-[System.CodeDom.Compiler.GeneratedCode("ANTLR", "4.7.2")]
+[System.CodeDom.Compiler.GeneratedCode("ANTLR", "4.13.1")]
+[System.Diagnostics.DebuggerNonUserCode]
 [System.CLSCompliant(false)]
 public partial class FBTBaseListener : IFBTListener {
 	/// <summary>
@@ -106,66 +107,6 @@ public partial class FBTBaseListener : IFBTListener {
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
 	public virtual void ExitTypeModule([NotNull] FBTParser.TypeModuleContext context) { }
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.classAttributes"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void EnterClassAttributes([NotNull] FBTParser.ClassAttributesContext context) { }
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.classAttributes"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void ExitClassAttributes([NotNull] FBTParser.ClassAttributesContext context) { }
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void EnterTypeAttributes([NotNull] FBTParser.TypeAttributesContext context) { }
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void ExitTypeAttributes([NotNull] FBTParser.TypeAttributesContext context) { }
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void EnterTypeAttribute([NotNull] FBTParser.TypeAttributeContext context) { }
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void ExitTypeAttribute([NotNull] FBTParser.TypeAttributeContext context) { }
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void EnterBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context) { }
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void ExitBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context) { }
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void EnterNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context) { }
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
-	/// <para>The default implementation does nothing.</para>
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	public virtual void ExitNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context) { }
 	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.typeEnum"/>.
 	/// <para>The default implementation does nothing.</para>
@@ -263,6 +204,150 @@ public partial class FBTBaseListener : IFBTListener {
 	/// <param name="context">The parse tree.</param>
 	public virtual void ExitClassInherrit([NotNull] FBTParser.ClassInherritContext context) { }
 	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeArray"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterTypeArray([NotNull] FBTParser.TypeArrayContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeArray"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitTypeArray([NotNull] FBTParser.TypeArrayContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.arrayTypeName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterArrayTypeName([NotNull] FBTParser.ArrayTypeNameContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.arrayTypeName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitArrayTypeName([NotNull] FBTParser.ArrayTypeNameContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.arrayRef"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterArrayRef([NotNull] FBTParser.ArrayRefContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.arrayRef"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitArrayRef([NotNull] FBTParser.ArrayRefContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.arrayRefName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterArrayRefName([NotNull] FBTParser.ArrayRefNameContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.arrayRefName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitArrayRefName([NotNull] FBTParser.ArrayRefNameContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeDelegate"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterTypeDelegate([NotNull] FBTParser.TypeDelegateContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeDelegate"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitTypeDelegate([NotNull] FBTParser.TypeDelegateContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeFunction"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterTypeFunction([NotNull] FBTParser.TypeFunctionContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeFunction"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitTypeFunction([NotNull] FBTParser.TypeFunctionContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionData"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterFunctionData([NotNull] FBTParser.FunctionDataContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionData"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitFunctionData([NotNull] FBTParser.FunctionDataContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionParamData"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterFunctionParamData([NotNull] FBTParser.FunctionParamDataContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionParamData"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitFunctionParamData([NotNull] FBTParser.FunctionParamDataContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionParamType"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterFunctionParamType([NotNull] FBTParser.FunctionParamTypeContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionParamType"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitFunctionParamType([NotNull] FBTParser.FunctionParamTypeContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionParamTypeName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterFunctionParamTypeName([NotNull] FBTParser.FunctionParamTypeNameContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionParamTypeName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitFunctionParamTypeName([NotNull] FBTParser.FunctionParamTypeNameContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionParamName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterFunctionParamName([NotNull] FBTParser.FunctionParamNameContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionParamName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitFunctionParamName([NotNull] FBTParser.FunctionParamNameContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeInterface"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterTypeInterface([NotNull] FBTParser.TypeInterfaceContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeInterface"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitTypeInterface([NotNull] FBTParser.TypeInterfaceContext context) { }
+	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.typeValueType"/>.
 	/// <para>The default implementation does nothing.</para>
 	/// </summary>
@@ -298,6 +383,42 @@ public partial class FBTBaseListener : IFBTListener {
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
 	public virtual void ExitValueTypeName([NotNull] FBTParser.ValueTypeNameContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.baseClassVariable"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterBaseClassVariable([NotNull] FBTParser.BaseClassVariableContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.baseClassVariable"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitBaseClassVariable([NotNull] FBTParser.BaseClassVariableContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.baseClassName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterBaseClassName([NotNull] FBTParser.BaseClassNameContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.baseClassName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitBaseClassName([NotNull] FBTParser.BaseClassNameContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.baseFieldName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterBaseFieldName([NotNull] FBTParser.BaseFieldNameContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.baseFieldName"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitBaseFieldName([NotNull] FBTParser.BaseFieldNameContext context) { }
 	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.structureData"/>.
 	/// <para>The default implementation does nothing.</para>
@@ -383,6 +504,126 @@ public partial class FBTBaseListener : IFBTListener {
 	/// <param name="context">The parse tree.</param>
 	public virtual void ExitMemberOffset([NotNull] FBTParser.MemberOffsetContext context) { }
 	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.memberDefaultValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterMemberDefaultValue([NotNull] FBTParser.MemberDefaultValueContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.memberDefaultValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitMemberDefaultValue([NotNull] FBTParser.MemberDefaultValueContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueInstance</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterMemberValueInstance([NotNull] FBTParser.MemberValueInstanceContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueInstance</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitMemberValueInstance([NotNull] FBTParser.MemberValueInstanceContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueArray</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterMemberValueArray([NotNull] FBTParser.MemberValueArrayContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueArray</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitMemberValueArray([NotNull] FBTParser.MemberValueArrayContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueNumber</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterMemberValueNumber([NotNull] FBTParser.MemberValueNumberContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueNumber</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitMemberValueNumber([NotNull] FBTParser.MemberValueNumberContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueFloat</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterMemberValueFloat([NotNull] FBTParser.MemberValueFloatContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueFloat</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitMemberValueFloat([NotNull] FBTParser.MemberValueFloatContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueString</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterMemberValueString([NotNull] FBTParser.MemberValueStringContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueString</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitMemberValueString([NotNull] FBTParser.MemberValueStringContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueNull</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterMemberValueNull([NotNull] FBTParser.MemberValueNullContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueNull</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitMemberValueNull([NotNull] FBTParser.MemberValueNullContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.memberInstanceField"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterMemberInstanceField([NotNull] FBTParser.MemberInstanceFieldContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.memberInstanceField"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitMemberInstanceField([NotNull] FBTParser.MemberInstanceFieldContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.memberArrayElement"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterMemberArrayElement([NotNull] FBTParser.MemberArrayElementContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.memberArrayElement"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitMemberArrayElement([NotNull] FBTParser.MemberArrayElementContext context) { }
+	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.typePrimitive"/>.
 	/// <para>The default implementation does nothing.</para>
 	/// </summary>
@@ -407,6 +648,90 @@ public partial class FBTBaseListener : IFBTListener {
 	/// <param name="context">The parse tree.</param>
 	public virtual void ExitPrimitiveTypeName([NotNull] FBTParser.PrimitiveTypeNameContext context) { }
 	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.classAttributes"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterClassAttributes([NotNull] FBTParser.ClassAttributesContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.classAttributes"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitClassAttributes([NotNull] FBTParser.ClassAttributesContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterTypeAttributes([NotNull] FBTParser.TypeAttributesContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitTypeAttributes([NotNull] FBTParser.TypeAttributesContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeAttributesData"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterTypeAttributesData([NotNull] FBTParser.TypeAttributesDataContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeAttributesData"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitTypeAttributesData([NotNull] FBTParser.TypeAttributesDataContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterTypeAttribute([NotNull] FBTParser.TypeAttributeContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitTypeAttribute([NotNull] FBTParser.TypeAttributeContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.stringTypeAttribute"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterStringTypeAttribute([NotNull] FBTParser.StringTypeAttributeContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.stringTypeAttribute"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitStringTypeAttribute([NotNull] FBTParser.StringTypeAttributeContext context) { }
+	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.numeral"/>.
 	/// <para>The default implementation does nothing.</para>
 	/// </summary>
@@ -418,6 +743,18 @@ public partial class FBTBaseListener : IFBTListener {
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
 	public virtual void ExitNumeral([NotNull] FBTParser.NumeralContext context) { }
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.float"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void EnterFloat([NotNull] FBTParser.FloatContext context) { }
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.float"/>.
+	/// <para>The default implementation does nothing.</para>
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	public virtual void ExitFloat([NotNull] FBTParser.FloatContext context) { }
 
 	/// <inheritdoc/>
 	/// <remarks>The default implementation does nothing.</remarks>
