@@ -1,14 +1,14 @@
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
-//     ANTLR Version: 4.7.2
+//     ANTLR Version: 4.13.1
 //
 //     Changes to this file may cause incorrect behavior and will be lost if
 //     the code is regenerated.
 // </auto-generated>
 //------------------------------------------------------------------------------
 
-// Generated from G:\Users\txt\Source\Repos\FBTC1\FBT.Generator\Parser\Grammar\FBT.g4 by ANTLR 4.7.2
+// Generated from FBT.g4 by ANTLR 4.13.1
 
 // Unreachable code detected
 #pragma warning disable 0162
@@ -27,7 +27,7 @@ using IToken = Antlr4.Runtime.IToken;
 /// This interface defines a complete listener for a parse tree produced by
 /// <see cref="FBTParser"/>.
 /// </summary>
-[System.CodeDom.Compiler.GeneratedCode("ANTLR", "4.7.2")]
+[System.CodeDom.Compiler.GeneratedCode("ANTLR", "4.13.1")]
 [System.CLSCompliant(false)]
 public interface IFBTListener : IParseTreeListener {
 	/// <summary>
@@ -90,56 +90,6 @@ public interface IFBTListener : IParseTreeListener {
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
 	void ExitTypeModule([NotNull] FBTParser.TypeModuleContext context);
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.classAttributes"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void EnterClassAttributes([NotNull] FBTParser.ClassAttributesContext context);
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.classAttributes"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void ExitClassAttributes([NotNull] FBTParser.ClassAttributesContext context);
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void EnterTypeAttributes([NotNull] FBTParser.TypeAttributesContext context);
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void ExitTypeAttributes([NotNull] FBTParser.TypeAttributesContext context);
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void EnterTypeAttribute([NotNull] FBTParser.TypeAttributeContext context);
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void ExitTypeAttribute([NotNull] FBTParser.TypeAttributeContext context);
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void EnterBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context);
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void ExitBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context);
-	/// <summary>
-	/// Enter a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void EnterNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context);
-	/// <summary>
-	/// Exit a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	void ExitNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context);
 	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.typeEnum"/>.
 	/// </summary>
@@ -221,6 +171,126 @@ public interface IFBTListener : IParseTreeListener {
 	/// <param name="context">The parse tree.</param>
 	void ExitClassInherrit([NotNull] FBTParser.ClassInherritContext context);
 	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeArray"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterTypeArray([NotNull] FBTParser.TypeArrayContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeArray"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitTypeArray([NotNull] FBTParser.TypeArrayContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.arrayTypeName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterArrayTypeName([NotNull] FBTParser.ArrayTypeNameContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.arrayTypeName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitArrayTypeName([NotNull] FBTParser.ArrayTypeNameContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.arrayRef"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterArrayRef([NotNull] FBTParser.ArrayRefContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.arrayRef"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitArrayRef([NotNull] FBTParser.ArrayRefContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.arrayRefName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterArrayRefName([NotNull] FBTParser.ArrayRefNameContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.arrayRefName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitArrayRefName([NotNull] FBTParser.ArrayRefNameContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeDelegate"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterTypeDelegate([NotNull] FBTParser.TypeDelegateContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeDelegate"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitTypeDelegate([NotNull] FBTParser.TypeDelegateContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeFunction"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterTypeFunction([NotNull] FBTParser.TypeFunctionContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeFunction"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitTypeFunction([NotNull] FBTParser.TypeFunctionContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionData"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterFunctionData([NotNull] FBTParser.FunctionDataContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionData"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitFunctionData([NotNull] FBTParser.FunctionDataContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionParamData"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterFunctionParamData([NotNull] FBTParser.FunctionParamDataContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionParamData"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitFunctionParamData([NotNull] FBTParser.FunctionParamDataContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionParamType"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterFunctionParamType([NotNull] FBTParser.FunctionParamTypeContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionParamType"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitFunctionParamType([NotNull] FBTParser.FunctionParamTypeContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionParamTypeName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterFunctionParamTypeName([NotNull] FBTParser.FunctionParamTypeNameContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionParamTypeName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitFunctionParamTypeName([NotNull] FBTParser.FunctionParamTypeNameContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.functionParamName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterFunctionParamName([NotNull] FBTParser.FunctionParamNameContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.functionParamName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitFunctionParamName([NotNull] FBTParser.FunctionParamNameContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeInterface"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterTypeInterface([NotNull] FBTParser.TypeInterfaceContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeInterface"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitTypeInterface([NotNull] FBTParser.TypeInterfaceContext context);
+	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.typeValueType"/>.
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
@@ -250,6 +320,36 @@ public interface IFBTListener : IParseTreeListener {
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
 	void ExitValueTypeName([NotNull] FBTParser.ValueTypeNameContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.baseClassVariable"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterBaseClassVariable([NotNull] FBTParser.BaseClassVariableContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.baseClassVariable"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitBaseClassVariable([NotNull] FBTParser.BaseClassVariableContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.baseClassName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterBaseClassName([NotNull] FBTParser.BaseClassNameContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.baseClassName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitBaseClassName([NotNull] FBTParser.BaseClassNameContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.baseFieldName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterBaseFieldName([NotNull] FBTParser.BaseFieldNameContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.baseFieldName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitBaseFieldName([NotNull] FBTParser.BaseFieldNameContext context);
 	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.structureData"/>.
 	/// </summary>
@@ -321,6 +421,108 @@ public interface IFBTListener : IParseTreeListener {
 	/// <param name="context">The parse tree.</param>
 	void ExitMemberOffset([NotNull] FBTParser.MemberOffsetContext context);
 	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.memberDefaultValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterMemberDefaultValue([NotNull] FBTParser.MemberDefaultValueContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.memberDefaultValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitMemberDefaultValue([NotNull] FBTParser.MemberDefaultValueContext context);
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueInstance</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterMemberValueInstance([NotNull] FBTParser.MemberValueInstanceContext context);
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueInstance</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitMemberValueInstance([NotNull] FBTParser.MemberValueInstanceContext context);
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueArray</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterMemberValueArray([NotNull] FBTParser.MemberValueArrayContext context);
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueArray</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitMemberValueArray([NotNull] FBTParser.MemberValueArrayContext context);
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueNumber</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterMemberValueNumber([NotNull] FBTParser.MemberValueNumberContext context);
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueNumber</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitMemberValueNumber([NotNull] FBTParser.MemberValueNumberContext context);
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueFloat</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterMemberValueFloat([NotNull] FBTParser.MemberValueFloatContext context);
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueFloat</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitMemberValueFloat([NotNull] FBTParser.MemberValueFloatContext context);
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueString</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterMemberValueString([NotNull] FBTParser.MemberValueStringContext context);
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueString</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitMemberValueString([NotNull] FBTParser.MemberValueStringContext context);
+	/// <summary>
+	/// Enter a parse tree produced by the <c>MemberValueNull</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterMemberValueNull([NotNull] FBTParser.MemberValueNullContext context);
+	/// <summary>
+	/// Exit a parse tree produced by the <c>MemberValueNull</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitMemberValueNull([NotNull] FBTParser.MemberValueNullContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.memberInstanceField"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterMemberInstanceField([NotNull] FBTParser.MemberInstanceFieldContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.memberInstanceField"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitMemberInstanceField([NotNull] FBTParser.MemberInstanceFieldContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.memberArrayElement"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterMemberArrayElement([NotNull] FBTParser.MemberArrayElementContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.memberArrayElement"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitMemberArrayElement([NotNull] FBTParser.MemberArrayElementContext context);
+	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.typePrimitive"/>.
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
@@ -341,6 +543,76 @@ public interface IFBTListener : IParseTreeListener {
 	/// <param name="context">The parse tree.</param>
 	void ExitPrimitiveTypeName([NotNull] FBTParser.PrimitiveTypeNameContext context);
 	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.classAttributes"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterClassAttributes([NotNull] FBTParser.ClassAttributesContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.classAttributes"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitClassAttributes([NotNull] FBTParser.ClassAttributesContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterTypeAttributes([NotNull] FBTParser.TypeAttributesContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitTypeAttributes([NotNull] FBTParser.TypeAttributesContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeAttributesData"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterTypeAttributesData([NotNull] FBTParser.TypeAttributesDataContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeAttributesData"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitTypeAttributesData([NotNull] FBTParser.TypeAttributesDataContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterTypeAttribute([NotNull] FBTParser.TypeAttributeContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitTypeAttribute([NotNull] FBTParser.TypeAttributeContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.stringTypeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterStringTypeAttribute([NotNull] FBTParser.StringTypeAttributeContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.stringTypeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitStringTypeAttribute([NotNull] FBTParser.StringTypeAttributeContext context);
+	/// <summary>
 	/// Enter a parse tree produced by <see cref="FBTParser.numeral"/>.
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
@@ -350,4 +622,14 @@ public interface IFBTListener : IParseTreeListener {
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
 	void ExitNumeral([NotNull] FBTParser.NumeralContext context);
+	/// <summary>
+	/// Enter a parse tree produced by <see cref="FBTParser.float"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void EnterFloat([NotNull] FBTParser.FloatContext context);
+	/// <summary>
+	/// Exit a parse tree produced by <see cref="FBTParser.float"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	void ExitFloat([NotNull] FBTParser.FloatContext context);
 }
