@@ -1,14 +1,14 @@
 //------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
-//     ANTLR Version: 4.7.2
+//     ANTLR Version: 4.13.1
 //
 //     Changes to this file may cause incorrect behavior and will be lost if
 //     the code is regenerated.
 // </auto-generated>
 //------------------------------------------------------------------------------
 
-// Generated from G:\Users\txt\Source\Repos\FBTC1\FBT.Generator\Parser\Grammar\FBT.g4 by ANTLR 4.7.2
+// Generated from FBT.g4 by ANTLR 4.13.1
 
 // Unreachable code detected
 #pragma warning disable 0162
@@ -28,7 +28,7 @@ using IToken = Antlr4.Runtime.IToken;
 /// by <see cref="FBTParser"/>.
 /// </summary>
 /// <typeparam name="Result">The return type of the visit operation.</typeparam>
-[System.CodeDom.Compiler.GeneratedCode("ANTLR", "4.7.2")]
+[System.CodeDom.Compiler.GeneratedCode("ANTLR", "4.13.1")]
 [System.CLSCompliant(false)]
 public interface IFBTVisitor<Result> : IParseTreeVisitor<Result> {
 	/// <summary>
@@ -67,36 +67,6 @@ public interface IFBTVisitor<Result> : IParseTreeVisitor<Result> {
 	/// <param name="context">The parse tree.</param>
 	/// <return>The visitor result.</return>
 	Result VisitTypeModule([NotNull] FBTParser.TypeModuleContext context);
-	/// <summary>
-	/// Visit a parse tree produced by <see cref="FBTParser.classAttributes"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	/// <return>The visitor result.</return>
-	Result VisitClassAttributes([NotNull] FBTParser.ClassAttributesContext context);
-	/// <summary>
-	/// Visit a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	/// <return>The visitor result.</return>
-	Result VisitTypeAttributes([NotNull] FBTParser.TypeAttributesContext context);
-	/// <summary>
-	/// Visit a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	/// <return>The visitor result.</return>
-	Result VisitTypeAttribute([NotNull] FBTParser.TypeAttributeContext context);
-	/// <summary>
-	/// Visit a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	/// <return>The visitor result.</return>
-	Result VisitBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context);
-	/// <summary>
-	/// Visit a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
-	/// </summary>
-	/// <param name="context">The parse tree.</param>
-	/// <return>The visitor result.</return>
-	Result VisitNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context);
 	/// <summary>
 	/// Visit a parse tree produced by <see cref="FBTParser.typeEnum"/>.
 	/// </summary>
@@ -146,6 +116,78 @@ public interface IFBTVisitor<Result> : IParseTreeVisitor<Result> {
 	/// <return>The visitor result.</return>
 	Result VisitClassInherrit([NotNull] FBTParser.ClassInherritContext context);
 	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.typeArray"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitTypeArray([NotNull] FBTParser.TypeArrayContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.arrayTypeName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitArrayTypeName([NotNull] FBTParser.ArrayTypeNameContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.arrayRef"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitArrayRef([NotNull] FBTParser.ArrayRefContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.arrayRefName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitArrayRefName([NotNull] FBTParser.ArrayRefNameContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.typeDelegate"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitTypeDelegate([NotNull] FBTParser.TypeDelegateContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.typeFunction"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitTypeFunction([NotNull] FBTParser.TypeFunctionContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.functionData"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitFunctionData([NotNull] FBTParser.FunctionDataContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.functionParamData"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitFunctionParamData([NotNull] FBTParser.FunctionParamDataContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.functionParamType"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitFunctionParamType([NotNull] FBTParser.FunctionParamTypeContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.functionParamTypeName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitFunctionParamTypeName([NotNull] FBTParser.FunctionParamTypeNameContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.functionParamName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitFunctionParamName([NotNull] FBTParser.FunctionParamNameContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.typeInterface"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitTypeInterface([NotNull] FBTParser.TypeInterfaceContext context);
+	/// <summary>
 	/// Visit a parse tree produced by <see cref="FBTParser.typeValueType"/>.
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
@@ -163,6 +205,24 @@ public interface IFBTVisitor<Result> : IParseTreeVisitor<Result> {
 	/// <param name="context">The parse tree.</param>
 	/// <return>The visitor result.</return>
 	Result VisitValueTypeName([NotNull] FBTParser.ValueTypeNameContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.baseClassVariable"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitBaseClassVariable([NotNull] FBTParser.BaseClassVariableContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.baseClassName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitBaseClassName([NotNull] FBTParser.BaseClassNameContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.baseFieldName"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitBaseFieldName([NotNull] FBTParser.BaseFieldNameContext context);
 	/// <summary>
 	/// Visit a parse tree produced by <see cref="FBTParser.structureData"/>.
 	/// </summary>
@@ -206,6 +266,66 @@ public interface IFBTVisitor<Result> : IParseTreeVisitor<Result> {
 	/// <return>The visitor result.</return>
 	Result VisitMemberOffset([NotNull] FBTParser.MemberOffsetContext context);
 	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.memberDefaultValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitMemberDefaultValue([NotNull] FBTParser.MemberDefaultValueContext context);
+	/// <summary>
+	/// Visit a parse tree produced by the <c>MemberValueInstance</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitMemberValueInstance([NotNull] FBTParser.MemberValueInstanceContext context);
+	/// <summary>
+	/// Visit a parse tree produced by the <c>MemberValueArray</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitMemberValueArray([NotNull] FBTParser.MemberValueArrayContext context);
+	/// <summary>
+	/// Visit a parse tree produced by the <c>MemberValueNumber</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitMemberValueNumber([NotNull] FBTParser.MemberValueNumberContext context);
+	/// <summary>
+	/// Visit a parse tree produced by the <c>MemberValueFloat</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitMemberValueFloat([NotNull] FBTParser.MemberValueFloatContext context);
+	/// <summary>
+	/// Visit a parse tree produced by the <c>MemberValueString</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitMemberValueString([NotNull] FBTParser.MemberValueStringContext context);
+	/// <summary>
+	/// Visit a parse tree produced by the <c>MemberValueNull</c>
+	/// labeled alternative in <see cref="FBTParser.memberValue"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitMemberValueNull([NotNull] FBTParser.MemberValueNullContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.memberInstanceField"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitMemberInstanceField([NotNull] FBTParser.MemberInstanceFieldContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.memberArrayElement"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitMemberArrayElement([NotNull] FBTParser.MemberArrayElementContext context);
+	/// <summary>
 	/// Visit a parse tree produced by <see cref="FBTParser.typePrimitive"/>.
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
@@ -218,9 +338,57 @@ public interface IFBTVisitor<Result> : IParseTreeVisitor<Result> {
 	/// <return>The visitor result.</return>
 	Result VisitPrimitiveTypeName([NotNull] FBTParser.PrimitiveTypeNameContext context);
 	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.classAttributes"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitClassAttributes([NotNull] FBTParser.ClassAttributesContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.typeAttributes"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitTypeAttributes([NotNull] FBTParser.TypeAttributesContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.typeAttributesData"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitTypeAttributesData([NotNull] FBTParser.TypeAttributesDataContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.typeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitTypeAttribute([NotNull] FBTParser.TypeAttributeContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.basicTypeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitBasicTypeAttribute([NotNull] FBTParser.BasicTypeAttributeContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.numeralTypeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitNumeralTypeAttribute([NotNull] FBTParser.NumeralTypeAttributeContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.stringTypeAttribute"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitStringTypeAttribute([NotNull] FBTParser.StringTypeAttributeContext context);
+	/// <summary>
 	/// Visit a parse tree produced by <see cref="FBTParser.numeral"/>.
 	/// </summary>
 	/// <param name="context">The parse tree.</param>
 	/// <return>The visitor result.</return>
 	Result VisitNumeral([NotNull] FBTParser.NumeralContext context);
+	/// <summary>
+	/// Visit a parse tree produced by <see cref="FBTParser.float"/>.
+	/// </summary>
+	/// <param name="context">The parse tree.</param>
+	/// <return>The visitor result.</return>
+	Result VisitFloat([NotNull] FBTParser.FloatContext context);
 }
