@@ -1,18 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿namespace FBT.TypeData.Base.Attributes;
 
-namespace FBT.TypeData.Base.Attributes
+public class TypeNumeralAttribute
+	: TypeAttribute
 {
-    public class TypeNumeralAttribute
-        : TypeAttribute
-    {
-        public TypeNumeralAttribute( string p_Key, int p_Value )
-            : base( p_Key )
-        {
-            this.Value = p_Value;
-        }
+	public long Value;
 
-        public int Value;
-    }
+	public TypeNumeralAttribute(string p_Key, long p_Value)
+		: base(p_Key)
+	{
+		Value = p_Value;
+	}
 }
