@@ -2,7 +2,8 @@
 
 namespace FBT.TypeData.Member;
 
-public class TypeDataPrimitive
+public class TypeDataArray
 	: TypeDataBase
 {
+	public RefTypeData ArrayType { get; set; } = new();
 }
